* Source/CFRuntime.c: Do nothing if CFRetain/CFRelease are called on small objects. * Source/GSObjCRuntime.h: Cast to unsigned int to avoid warning
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/corebase/trunk@38351 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSObjCRuntime.h
@@ -35,8 +35,6 @@ GS_PRIVATE const void *CFTypeRetainCallBack (CFAllocatorRef alloc,
 
 GS_PRIVATE void CFTypeReleaseCallBack (CFAllocatorRef alloc, const void *value);
 
-
-
 #if HAVE_LIBOBJC || HAVE_LIBOBJC2
 
 #define BOOL OBJC_BOOL
@@ -57,10 +55,9 @@ CF_INLINE Boolean
 CF_IS_OBJC (CFTypeID typeID, const void *obj)
 {
 #if defined(OBJC_SMALL_OBJECT_MASK)
-  return (obj && ((obj & OBJC_SMALL_OBJECT_MASK) != 0
-                  || (typeID >= __CFRuntimeClassTableCount
-                      || object_getClass ((id) obj) !=
-                      __CFISAForTypeID (typeID))));
+  return (obj && (((unsigned long)obj & OBJC_SMALL_OBJECT_MASK) != 0
+                  || typeID >= __CFRuntimeClassTableCount
+                  || object_getClass ((id) obj) != __CFISAForTypeID (typeID)));
 #else
   return (obj && (typeID >= __CFRuntimeClassTableCount
                   || object_getClass ((id) obj) != __CFISAForTypeID (typeID)));
